1. Change some path settings of VS 2003 solution/project files to un-break the brokeback mountain... oops, the broken build process of Win32 version. But OVIMGeneric and OVIMGenericSQLite are still broken because of OVKeySequenceSimple.
2. Change proper int variables to size_t.

3. Add UTF-8 BOM to OVIMPOJ-Holo.cpp to make VC compiler happy.

+++ Experiments/Win32/AmphiVanilla/AVLoader.cpp
@@ -86,7 +86,7 @@ void AVLoader::initContext(int i)
 void AVLoader::reloadConfig()
 {
 	em->dict()->update();
-	for(int i = 0; i < em->modlist().size(); i++) {
+	for(size_t i = 0; i < em->modlist().size(); i++) {
 		if(ctx_vector.at(i) != NULL) {
 			murmur("Reload: %s", em->modlist().at(i)->identifier());
 			em->dict()->setDict(em->modlist().at(i)->identifier());
